+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **FisherIndicator**: commented constructor removed.
 - **TestUtils**: removed convenience methods for permuted parameters, fixed all unit tests
 - **BaseTimeSeries**: _Constructor_ `BaseTimeSeries(TimeSeries defaultSeries, int seriesBeginIndex, int seriesEndIndex)` _removed_. Use `TimeSeries.getSubseries(int i, int i)` instead
+- **BigDecimalNum**: _removed_.  Replaced by `PrecisionNum`
 
 ## 0.11 (released January 25, 2018)